[DD-44] Add Restaurant image
Added restaurant image

+++ backend/controllers/photoController.js
@@ -1,8 +1,6 @@
-const fs = require('fs');
-const path = require('path');
-
 var PhotoModel = require('../models/photoModel.js');
 var UserModel = require('../models/userModel.js');
+var RestaurantModel = require('../models/restaurantModel.js');
 
 /**
  * photoController.js
@@ -54,9 +52,8 @@ module.exports = {
     /**
      * photoController.create()
      */
-    create: async function (req, res) {
+    createProfilePhoto: async function (req, res) {
         try {
-
             var newPhoto = new PhotoModel({
                 imagePath : "/images/" + req.file.filename
             });
@@ -67,15 +64,33 @@ module.exports = {
 
             if(user.profilePhoto.toString() !== process.env.DEFAULT_AVATAR_ID) {
                 const oldPhoto = await PhotoModel.findOneAndDelete({_id: user.profilePhoto._id});
-
-                fs.unlink(path.join(__dirname, '..', 'public', oldPhoto.imagePath), err => {
-                    if (err) {
-                        console.error('Error when deleting the photo file:', err);
-                    }
-                });
             }
             user.profilePhoto = newPhoto._id;
             await user.save();
+            res.status(200).json(newPhoto);
+        } catch (err) {
+            return res.status(500).json({
+                message: 'Error when creating photo',
+                error: err
+            });
+        }
+    },
+
+    createRestaurantPhoto: async function (req, res) {
+        try {
+            var newPhoto = new PhotoModel({
+                imagePath : "/images/" + req.file.filename
+            });
+
+            newPhoto = await newPhoto.save();
+
+            const restaurant = await RestaurantModel.findOne({_id: req.params.restaurantId});
+
+            if(restaurant.photo.toString() !== process.env.DEFAULT_RESTAURANT_PHOTO_ID) {
+                const oldPhoto = await PhotoModel.findOneAndDelete({_id: restaurant.photo});
+            }
+            restaurant.photo = newPhoto._id;
+            await restaurant.save();
             res.status(200).json(newPhoto);
         } catch (err) {
             return res.status(500).json({
@@ -123,10 +138,41 @@ module.exports = {
     /**
      * photoController.remove()
      */
-    remove: function (req, res) {
+    removeProfilePhoto: function (req, res) {
         var id = req.params.id;
 
-        if(id !== process.env.DEFAULT_AVATAR_ID) {
+        if(id !== process.env.DEFAULT_AVATAR_ID && id !== process.env.DEFAULT_RESTAURANT_PHOTO_ID) {
+            PhotoModel.findOneAndDelete({ _id: id })
+            .then(photo => {
+                if (!photo) {
+                    return res.status(404).json({
+                        message: 'No such photo'
+                    });
+                }
+                UserModel.findOneAndUpdate({ profilePhoto: id }, { profilePhoto: process.env.DEFAULT_AVATAR_ID }, { new: true })
+                .then(() => {
+                    return res.status(204).json();
+                })
+                .catch(err => {
+                    return res.status(500).json({
+                        message: "Error when updating user's profile photo after deletion",
+                        error: err
+                    });
+                });
+            })
+            .catch(err => {
+                return res.status(500).json({
+                    message: 'Error when deleting the photo.',
+                    error: err
+                });
+            })
+        }
+    },
+
+    removeRestaurantPhoto: function (req, res) {
+        var id = req.params.id;
+
+        if(id !== process.env.DEFAULT_AVATAR_ID && id !== process.env.DEFAULT_RESTAURANT_PHOTO_ID) {
             PhotoModel.findOneAndDelete({ _id: id })
             .then(photo => {
                 if (!photo) {
@@ -135,19 +181,13 @@ module.exports = {
                     });
                 }
 
-                fs.unlink(path.join(__dirname, '..', 'public', photo.imagePath), err => {
-                    if (err) {
-                        console.error('Error when deleting the photo file:', err);
-                    }
-                });
-
-                UserModel.findOneAndUpdate({ profilePhoto: id }, { profilePhoto: process.env.DEFAULT_AVATAR_ID }, { new: true })
+                RestaurantModel.findOneAndUpdate({ photo: id }, { photo: process.env.DEFAULT_RESTAURANT_PHOTO_ID }, { new: true })
                 .then(() => {
                     return res.status(204).json();
                 })
                 .catch(err => {
                     return res.status(500).json({
-                        message: "Error when updating user's profile photo after deletion",
+                        message: "Error when updating restaurants's photo after deletion",
                         error: err
                     });
                 });
